Add files via upload
There's the projet about the link between good habits, and income on NHIS Data (US Health data survey). 
The method is panel datas, and I try to manipulate correctly the database, to face with the challenges of the encoding variables , missing data, big datas. 
Too, I tried to try a within method, which controls from the fixed unobserval, and that helps to have a more causal effect ! 
But I will retry with  a better adequation of the data , and find solutions to controls the validity of the model in a new papers. 
It's only a try.